requirements.txt
git+https://github.com/Rapptz/discord.py@45d498c1#[voice]
ruamel.yaml
Pillow
PyNaCl
gTTS
sox
